refactor: move input style types and default colors into platform-neutral modules
