+++ packages/react-native-enriched-markdown/src/inputStyleDefaults.ts
@@ -1,0 +1,6 @@
+// Default input colors shared by the native normalizer and the web editor.
+
+export const DEFAULT_LINK_COLOR = '#2563EB';
+export const DEFAULT_LINK_BG_COLOR = 'transparent';
+export const DEFAULT_SPOILER_COLOR = '#374151';
+export const DEFAULT_SPOILER_BG_COLOR = '#E5E7EB';
